requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
jira>=3.10.0
pydantic>=2.0
pydantic-settings>=2.0
mcp>=1.7.0
python-dotenv>=1.0.0
